+++ config/noro.json
@@ -1,5 +1,5 @@
 {
-    "base_config": "config/tts.json",
+    "base_config": "config/base.json",
     "model_type": "VC",
     "dataset": ["mls"],
     "model": {
@@ -18,13 +18,12 @@
             "num_query_emb": 32
         },
         "diffusion": {
-            "diffusion_type": "diffusion",   // Option: "diffusion" or "flow matching"
             "beta_min": 0.05,
             "beta_max": 20,
             "sigma": 1.0,
             "noise_factor": 1.0,
             "ode_solve_method": "euler",
-            "diff_model_type": "WaveNet",   // Transformer or WaveNet
+            "diff_model_type": "WaveNet",   
             "diff_wavenet":{
                 "input_size": 80,
                 "hidden_size": 512,
