+++ egs/vc/Noro/exp_config_clean.json
@@ -3,23 +3,15 @@
   "dataset": [
     "mls"
   ],
-  "preprocess": {
-    // Specify the output root path to save the processed data 
-    "processed_dir": "data",
-    "train_file": "train.json",
-    "valid_file": "test.json",
-    "read_metadata": true,
-    "metadata_dir": "metadata"
-  },
   // Specify the output root path to save model checkpoints and logs
   "log_dir": "path/to/your/checkpoint/directory",
   "train": {
     // New trainer and Accelerator
     "gradient_accumulation_step": 1,
     "tracker": ["tensorboard"],
-    "max_epoch": 5,
+    "max_epoch": 10,
     "save_checkpoint_stride": [1000],
-    "keep_last": [100],
+    "keep_last": [20],
     "run_eval": [true],
     "dataloader": {
       "num_worker": 64,
@@ -29,19 +21,22 @@
       "lr": 5e-5
     },
     "use_dynamic_batchsize": true,
-    "batch_size": 8,
-    "max_tokens": 800000,
+    "max_tokens": 3200000,
     "max_sentences": 64,
     "lr_warmup_steps": 5000,
     "lr_scheduler": "cosine",
     "num_train_steps": 800000
   },
   "trans_exp": {
-    "dataset_list": ["librilight_large", "librilight_medium", "librilight_small"],
-    "test_dataset_list": ["mls_test"],
-    "cache_dir": "path/to/your/exp_cache/directory",
-    "use_speaker": true,
-    "use_noise": true,
+    "directory_list": [
+      "path/to/your/training_data_directory1",
+      "path/to/your/training_data_directory2",
+      "path/to/your/training_data_directory3"
+    ],
+    "use_ref_noise": false,
+    "snr_lower": 2,
+    "snr_upper": 42,
+    "total_snrlevels": 5,
     "noise_dir": "path/to/your/noise/train/directory",
     "test_noise_dir": "path/to/your/noise/test/directory"
   }
