--- egs/vc/exp_config_base.json
@@ -1,101 +1,0 @@
-{
-    "base_config": "config/vc.json",
-    "model_type": "VC",
-    "dataset": [
-        "mls"
-    ],
-    "preprocess": {
-        "use_mel": false,
-        "use_code": true,
-        "use_spkid": true,
-        "use_pitch": true,
-        "use_duration": true,
-        "use_phone": true,
-        "use_len": true,
-        "use_cross_reference": true,
-        "train_file": "train.json",
-        "valid_file": "test.json",
-        "melspec_dir": "mel",
-        "code_dir": "code",
-        "pitch_dir": "pitch",
-        "duration_dir": "duration",
-        "metadata_dir": "metadata",
-        "read_metadata": true,
-        "clip_mode": "start"
-    },
-    "model": {
-        "reference_encoder": {
-            "encoder_layer": 6,
-            "encoder_hidden": 512,
-            "encoder_head": 8,
-            "conv_filter_size": 2048,
-            "conv_kernel_size": 9,
-            "encoder_dropout": 0.2,
-            "use_skip_connection": false,
-            "use_new_ffn": true,
-            "ref_in_dim": 80,
-            // "ref_in_dim": 96,
-            "ref_out_dim": 512,
-            "use_query_emb": true,
-            "num_query_emb": 32
-        },
-        "diffusion": {
-            "diffusion_type": "diffusion",   // Option: "diffusion" or "flow matching"
-            "beta_min": 0.05,
-            "beta_max": 20,
-            "sigma": 1.0,
-            "noise_factor": 1.0,
-            "ode_solve_method": "euler",
-            "diff_model_type": "WaveNet",   // Transformer or WaveNet
-            "diff_wavenet":{
-                // "input_size": 96,
-                "input_size": 80,
-                "hidden_size": 512,
-                "out_size": 80,
-                // "out_size": 96,
-                "num_layers": 47,
-                "cross_attn_per_layer": 3,
-                "dilation_cycle": 2,
-                "attn_head": 8,
-                "drop_out": 0.2
-            }
-        },
-        "prior_encoder": {
-            "encoder_layer": 6,
-            "encoder_hidden": 512,
-            "encoder_head": 8,
-            "conv_filter_size": 2048,
-            "conv_kernel_size": 9,
-            "encoder_dropout": 0.2,
-            "use_skip_connection": false,
-            "use_new_ffn": true,
-            "vocab_size": 256,
-            "cond_dim": 512,
-            "duration_predictor": {
-                "input_size": 512,
-                "filter_size": 512,
-                "kernel_size": 3,
-                "conv_layers": 30,
-                "cross_attn_per_layer": 3,
-                "attn_head": 8,
-                "drop_out": 0.2
-            },
-            "pitch_predictor": {
-                "input_size": 512,
-                "filter_size": 512,
-                "kernel_size": 5,
-                "conv_layers": 30,
-                "cross_attn_per_layer": 3,
-                "attn_head": 8,
-                "drop_out": 0.5
-            },
-            "pitch_min": 50,
-            "pitch_max": 1100,
-            "pitch_bins_num": 512
-        },
-        "vc_feature": {
-            "content_feature_dim": 768,
-            "hidden_dim": 512
-        }
-    }
-}
