+++ models/vc/Noro/noro_inference.py
@@ -1,3 +1,8 @@
+# Copyright (c) 2023 Amphion.
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
 import argparse
 import torch
 import numpy as np
@@ -5,14 +10,15 @@ import librosa
 from safetensors.torch import load_model
 import os
 from utils.util import load_config
-from models.vc.vc_trainer import VCTrainer
-from models.vc.ns2_uniamphion import UniAmphionVC
-from models.vc.hubert_kmeans import HubertWithKmeans
-from models.vc.vc_utils import mel_spectrogram, extract_world_f0
+from models.vc.Noro.noro_trainer import NoroTrainer
+from models.vc.Noro.noro_model import Noro_VCmodel
+from processors.content_extractor import HubertExtractor
+from utils.mel import mel_spectrogram_torch
+from utils.f0 import get_f0_features_using_dio, interpolate
 
 def build_trainer(args, cfg):
     supported_trainer = {
-        "VC": VCTrainer,
+        "VC": NoroTrainer,
     }
     trainer_class = supported_trainer[cfg.model_type]
     trainer = trainer_class(args, cfg)
@@ -70,11 +76,11 @@ def main():
         torch.cuda.empty_cache()
     ckpt_path = args.checkpoint_path
     
-    w2v = HubertWithKmeans()
+    w2v = HubertExtractor(cfg)
     w2v = w2v.to(device=args.local_rank)
     w2v.eval()
 
-    model = UniAmphionVC(cfg=cfg.model)
+    model = Noro_VCmodel(cfg=cfg.model)
     print("Loading model")
 
     load_model(model, ckpt_path)
@@ -96,14 +102,18 @@ def main():
     ref_audio = ref_audio[None, :]
     
     with torch.no_grad():
-        ref_mel = mel_spectrogram(ref_audio)
+        ref_mel = mel_spectrogram_torch(ref_audio, cfg)
         ref_mel = ref_mel.transpose(1, 2).to(device=args.local_rank)
         ref_mask = torch.ones(ref_mel.shape[0], ref_mel.shape[1]).to(args.local_rank).bool()
 
-        _, content_feature = w2v(audio)
+        _, content_feature = w2v.extract_content_features(audio)
         content_feature = content_feature.to(device=args.local_rank)
 
-        pitch_raw = extract_world_f0(audio)
+        wav = audio.cpu().numpy()
+        pitch_raw = get_f0_features_using_dio(wav, cfg)
+        pitch_raw, _ = interpolate(pitch_raw)
+        frame_num = len(wav) // cfg.preprocess.hop_size
+        pitch_raw = torch.from_numpy(pitch_raw[:frame_num]).float()
         pitch = (pitch_raw - pitch_raw.mean(dim=1, keepdim=True)) / (
             pitch_raw.std(dim=1, keepdim=True) + 1e-6
         )
